requirements.txt
Numpy
pandas
pymysql
python-dotenv
scikit-learn